Time: 0 ms (100.00%) | Memory: 40.2 MB (36.05%) - LeetSync

+++ 2608-count-the-digits-that-divide-a-number/count-the-digits-that-divide-a-number.java
@@ -5,7 +5,7 @@ class Solution {
         for(int i=0; i<val.length(); i++){
 
       
-            if(num%((int)val.charAt(i)-'0')==0)count++;
+            if(num%(val.charAt(i)-'0')==0)count++;
         }
         return count;
     }
